+++ docs/conf.py
@@ -22,7 +22,7 @@ extensions = [
 
 # templates_path = ['_templates']
 exclude_patterns = [".venv", "_build", "Thumbs.db", ".DS_Store"]
-
+add_module_names = False
 
 # -- Options for HTML output --
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output